(#10) Subsequent comment lines should align with the first one
Fixes #10.

I also added a new test file that could be used in cases like this, when there
is no way to reproduce the error within the usual test_indent.erl file. (In
this case, the new test relies on the fact the only the second comment line is
indented by the test, the first one must be left as it is.)