syntax: Fix closing docstring with ".
Previously the script highlighted the following incorrectly:

    -moduledoc """
    A module for basic "arithmetic".
    """.

It thought that the `".` after the word "arithmetic" closed the
docstring and the next line opened a new docstring.

This was because in Vim, the later rules are stronger, so the rule for
"..." beat the rule for """...""".

This commit fixes the problem by moving the rules for """...""" later.